Implementacion de componetes y mejoras en el codigo

+++ database/migrations/2024_03_05_051608_create_investments_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('investments', function (Blueprint $table) {
             $table->id();
             $table->string('type',50);
-            $table->float('amount');
+            $table->decimal('amount',10,2);
             $table->date('investmentDate');
             $table->string('return');
             $table->enum('status',['En curso', 'Finalizado']);
